Added page banks, backlight toggle, more characters

+++ software/buttons.py
@@ -9,7 +9,16 @@ mouse = Controller()
 
 import pyautogui
 
-page = 6
+page = 5
+bank = 1
+
+backlight = True
+
+import RPi.GPIO as GPIO
+
+GPIO.setmode(GPIO.BOARD)
+GPIO.setwarnings(False)
+GPIO.setup(13,GPIO.OUT)
 
 buttons = [False,False,False,False,False,False,False,False,False]
 
@@ -43,46 +52,80 @@ def menuHan(channel, button):
 def button_callback1(channel):
     motion = menuHan(channel, 1)
     
+    global bank
     global page
     
     if motion == "down":
-        if page is 1:
-            pyautogui.press('q')
+        if bank is 1:
+            if page is 1:
+                pyautogui.press('q')
+                
+            elif page is 2:
+                pyautogui.press('b')
             
-        elif page is 2:
-            pyautogui.press('b')
-        
-        elif page is 3:
-            pyautogui.press('d')
-        
-        elif page is 4:
-            pyautogui.press('s')
-        
-        elif page is 8:
-            os.system('notify-send "blocked mode"')
-            sleep(10)
-            os.system('pkill dis.sh')
-            os.system('pkill psp.sh')
-			sleep(1)
-            os.system('pkill python3')
+            elif page is 3:
+                pyautogui.press('d')
+            
+            elif page is 4:
+                pyautogui.press('s')
+                
+            elif page is 6:
+                pyautogui.press('1')
+                
+            elif page is 7:
+                pyautogui.press('0')
+            
+            elif page is 8:
+                os.system('notify-send "blocked mode"')
+                sleep(10)
+                os.system('pkill dis.sh')
+                os.system('pkill psp.sh')
+                sleep(1)
+                os.system('pkill python3')
+				print("backlight off")
+				GPIO.output(13,0)
+				backlight = False
     
 def button_callback2(channel):
     motion = menuHan(channel, 2)
     
+    global bank
     global page
     
+    global backlight
+    
     if motion == "down":
-        if page is 1:
-            pyautogui.press('u')
+        if bank is 1:
+            if page is 1:
+                pyautogui.press('u')
+                
+            elif page is 2:
+                pyautogui.press('r')
             
-        elif page is 2:
-            pyautogui.press('r')
-        
-        elif page is 3:
-            pyautogui.press('j')
-        
-        elif page is 4:
-            pyautogui.press('v')
+            elif page is 3:
+                pyautogui.press('j')
+            
+            elif page is 4:
+                pyautogui.press('v')
+                
+            elif page is 5:
+                pyautogui.press('up')
+                
+            elif page is 6:
+                pyautogui.press('2')
+                
+            elif page is 7:
+                pyautogui.press('3')
+                
+            elif page is 8:
+                if backlight:
+                    print("backlight off")
+                    GPIO.output(13,0)
+                    backlight = False
+                else:
+                    print("backlight on")
+                    GPIO.output(13,1)
+                    backlight = True
     
 def button_callback3(channel):
     print("Button 3")
@@ -102,104 +145,184 @@ def button_callback3(channel):
 def button_callback4(channel):
     motion = menuHan(channel, 4)
     
+    global bank
     global page
     
     if motion == "down":
-        if page is 1:
-            pyautogui.press('i')
+        if bank is 1:
+            if page is 1:
+                pyautogui.press('i')
+                
+            elif page is 2:
+                pyautogui.press('o')
             
-        elif page is 2:
-            pyautogui.press('o')
-        
-        elif page is 3:
-            pyautogui.press('h')
+            elif page is 3:
+                pyautogui.press('h')
+                
+            elif page is 4:
+                pyautogui.press(' ')
+                
+            elif page is 5:
+                pyautogui.press('left')
+                
+            elif page is 6:
+                pyautogui.press('4')
+                
+            elif page is 7:
+                pyautogui.press('$')
             
 def button_callback5(channel):
     motion = menuHan(channel, 5)
     
+    global bank
     global page
     
     if motion == "down":
-        if page is 1:
-            pyautogui.press('c')
+        if bank is 1:
+            if page is 1:
+                pyautogui.press('c')
+                
+            elif page is 2:
+                pyautogui.press('w')
             
-        elif page is 2:
-            pyautogui.press('w')
-        
-        elif page is 3:
-            pyautogui.press('m')
+            elif page is 3:
+                pyautogui.press('m')
+                
+            elif page is 4:
+                pyautogui.press('backspace')
+                
+            elif page is 6:
+                pyautogui.press('5')
+                
+            elif page is 7:
+                pyautogui.press('#')
 
     
 def button_callback6(channel):
     motion = menuHan(channel, 7)
     
+    global bank
     global page
     
     if motion == "down":
-        if page is 1:
-            pyautogui.press('g')
+        if bank is 1:
+            if page is 1:
+                pyautogui.press('g')
+                
+            elif page is 2:
+                pyautogui.press('f')
             
-        elif page is 2:
-            pyautogui.press('f')
-        
-        elif page is 3:
-            pyautogui.press('l')
-            
-    if page is 6:
-        if motion == "up":
-            print("Mouse up")
-            mouse.release(Button.left)
-        else:
-            print("Mouse down")
-            mouse.press(Button.left)
+            elif page is 3:
+                pyautogui.press('l')
+                
+            elif page is 4:
+                pyautogui.press('del')
+                
+            elif page is 6:
+                pyautogui.press('7')
+    
+    if bank is 1:
+        if page is 5:
+            if motion == "up":
+                print("Mouse up")
+                mouse.release(Button.left)
+            else:
+                print("Mouse down")
+                mouse.press(Button.left)
 
 def button_callback7(channel):
     motion = menuHan(channel, 6)
     
+    global bank
+    global page
+    
     if motion == "down":
-        if page is 1:
-            pyautogui.press('k')
+        if bank is 1:
+            if page is 1:
+                pyautogui.press('k')
+                
+            elif page is 2:
+                pyautogui.press('n')
             
-        elif page is 2:
-            pyautogui.press('n')
-        
-        elif page is 3:
-            pyautogui.press('p')
+            elif page is 3:
+                pyautogui.press('p')
+                
+            elif page is 4:
+                pyautogui.press('enter')
+                
+            elif page is 5:
+                pyautogui.press('right')
+                
+            elif page is 6:
+                pyautogui.press('6')
+                
+            elif page is 7:
+                pyautogui.press('|')
             
     
 def button_callback8(channel):
     motion = menuHan(channel, 8)
     
+    global bank
+    global page
+    
     if motion == "down":
-        if page is 1:
-            pyautogui.press('e')
+        if page is 8:
+            bank = 2
             
-        elif page is 2:
-            pyautogui.press('a')
-        
-        elif page is 3:
-            pyautogui.press('z')
+        if bank is 1:
+            if page is 1:
+                pyautogui.press('e')
+                
+            elif page is 2:
+                pyautogui.press('a')
+            
+            elif page is 3:
+                pyautogui.press('z')
+            
+            elif page is 4:
+                pyautogui.press('capslock')
+                
+            elif page is 5:
+                pyautogui.press('down')
+                
+            elif page is 6:
+                pyautogui.press('8')
     
 def button_callback9(channel):
     motion = menuHan(channel, 9)
-
-    if motion == "down":
-        if page is 1:
-            pyautogui.press('t')
-            
-        elif page is 2:
-            pyautogui.press('x')
-        
-        elif page is 3:
-            pyautogui.press('y')
     
-    if page is 6:
-        if motion == "up":
-            print("RMB Mouse up")
-            mouse.release(Button.right)
-        else:
-            print("RMB Mouse down")
-            mouse.press(Button.right)
+    global bank
+    global page
+    
+    if motion == "down":
+        if page is 8:
+            bank = 1
+    
+        if bank is 1:
+            if page is 1:
+                pyautogui.press('t')
+                
+            elif page is 2:
+                pyautogui.press('x')
+            
+            elif page is 3:
+                pyautogui.press('y')
+                
+            elif page is 4:
+                pyautogui.press('tab')
+                
+            elif page is 6:
+                pyautogui.press('9')
+    
+    if bank is 1:
+        if page is 5:
+            if motion == "up":
+                print("RMB Mouse up")
+                mouse.release(Button.right)
+            else:
+                print("RMB Mouse down")
+                mouse.press(Button.right)
             
             
 """
@@ -209,6 +332,9 @@ _ = nothing
 
 [M] = Page selector - hold and press another button to select page
 
+[B#] = Bank selector (changes to bank #) - press to change the bank of pages in use to the one specified
+
+backspace
 
 page selector:
 
@@ -246,25 +372,52 @@ l z y
 
 s v [M]
 
-_ _ _
+space backspace enter
 
-_ _ _
+del capslock tab
+
+
+---page 5---
+
+_ up [M]
+
+left _ right
+
+LMB down RMB
+
+---page 6---
+
+1 2 [M]
+
+4 5 6
+
+7 8 9
 
 ---page 7---
 
-_ _ _
+0 3 [M]
+
+$ # |
 
 _ _ _
 
-LMB _ RMB
+---page 8---
 
----page 9---
-
-[kill] _ _
+[kill] backlight-toggle [M]
 
 _ _ _
 
-_ _ _
+_ [B2] [B1]
+
+
+---TO ADD---
+
+'!', '"', '%', '&', "'", '(',
+')', '*', '+', ',', '-', '.', '/', 
+':', ';', '<', '=', '>', '?', '@', '[', '\\', ']', '^', '_', '`', 
+'{', '}', '~',
+'alt', 'ctrl', 'end', 'esc', 'pagedown', 'pageup', 'prntscrn',
+'shift', 'volumedown', 'volumemute', 'volumeup', 'win',
 
 """
